// src/app/api-client/auth/model/schoolRegistrationDTO.ts
/**
 * TeachingBACKEND API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SchoolRegistrationDTO { 
    firstName: string;
    password: string;
    email: string;
    phoneNumber: string;
    profession: string;
    city: string;
    postalCode: string;
    school: string;
}